fixed net message streams tests
